Updated target implementation and added alias constants `TargetJS` and `TargetTS`
