+++ kipper/target-ts/README.md
@@ -25,6 +25,54 @@ npm i @kipper/target-ts
 
 If you are using `pnpm` or `yarn`, use `pnpm i @kipper/target-ts` or `yarn add @kipper/target-ts`.
 
+## Usage
+
+Simply import the target and specify it in the `compilerOptions` field of `KipperCompiler.compile()`, for example:
+
+- JavaScript (CommonJS):
+
+	```js
+	const fs = require("fs").promises;
+	const kipper = require("@kipper/core");
+	const kipperTS = require("@kipper/target-ts");
+
+	const path = "INSERT_PATH";
+	fs.readFile(path, "utf8").then(async (fileContent) => {
+		const compiler = new kipper.KipperCompiler();
+
+		// Compile the code string or stream
+		let result = await compiler.compile(fileContent, { target: new kipperTS.TargetTS() });
+		let jsCode = result.write();
+
+		// Running the Kipper program
+		eval(jsCode);
+	});
+	```
+
+- TypeScript (CommonJS):
+
+	```ts
+	import { promises as fs } from "fs";
+	import { KipperCompiler } from "@kipper/core";
+	import TargetTS from "@kipper/target-ts";
+
+	const path = "INSERT_PATH";
+	fs.readFile(path, "utf8" as BufferEncoding).then(async (fileContent: string) => {
+		const compiler = new KipperCompiler();
+
+		// Compile the code string or stream
+		let result = await compiler.compile(fileContent, { target: new TargetTS() });
+		let tsCode = result.write();
+	});
+	```
+ 
+If you are using `@kipper/cli` then this package is automatically installed and compiling to TypeScript can be done
+using the `--target=ts` flag, for example:
+
+```bash
+kipper compile example-script.kip --target=ts
+```
+
 ## Kipper Docs
 
 Proper documentation for the Kipper language is available [here](https://luna-klatzer.github.io/Kipper/)!
